Settle on weak but objective log-based definition of "group chat"
